Update Angular contact demo with working Proof of Concept
Angular must add the custom event listener, and due to
scope and access of DOM elements, this must occur
outside of the Angular component context. Really, this
script should be rewritten and the contact form be broken
down into a series of components & directives that
achieve the same thing in a safer and non-globally
affected way.